+++ isaaclab_arena/environment_spec/arena_env_graph_spec.py
@@ -16,6 +16,7 @@ from isaaclab_arena.environment_spec.arena_env_graph_types import (
     CliOverrideSpec,
     CompositeTaskSpec,
     ObjectReferenceSpec,
+    PlacementSpec,
     SpatialRelationSpec,
     TaskSpec,
 )
@@ -39,6 +40,10 @@ class ArenaEnvGraphSpec(BaseModel):
     )
     relations: list[SpatialRelationSpec] = Field(
         default_factory=list, description="Spatial layout relations across all assets."
+    )
+    placement: PlacementSpec | None = Field(
+        default=None,
+        description="Per-env object-placement validators; null runs all build-time checks.",
     )
     task: CompositeTaskSpec = Field(description="Root task the robot performs to manipulate the objects.")
     cli_override_specs: list[CliOverrideSpec] | None = Field(
