+++ isaaclab_arena/relations/object_placer.py
@@ -6,52 +6,36 @@
 from __future__ import annotations
 
 import torch
-import trimesh
-from collections.abc import Iterator
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, TypeVar, cast
+from typing import TYPE_CHECKING
 
-from isaaclab_arena.relations.bounding_box_helpers import assign_variants_for_envs, build_per_env_bounding_boxes
-from isaaclab_arena.relations.collision_mode import CollisionMode, get_object_collision_mode, object_uses_mesh_collision
+from isaaclab_arena.relations.bounding_box_helpers import (
+    PerEnvBoundingBoxes,
+    assign_variants_for_envs,
+    build_per_env_bounding_boxes,
+)
 from isaaclab_arena.relations.object_placer_params import ObjectPlacerParams
 from isaaclab_arena.relations.placement_result import PlacementResult
-from isaaclab_arena.relations.placement_validation import PlacementCheck, PlacementValidationResults
-from isaaclab_arena.relations.relation_loss_strategies import (
-    SIDE_CONFIGS,
-    NotNextToLossStrategy,
-    next_to_violations,
-    not_next_to_violations,
-)
+from isaaclab_arena.relations.placement_validation import PlacementValidationResults
+from isaaclab_arena.relations.placement_validators import build_validators
 from isaaclab_arena.relations.relation_solver import RelationSolver
 from isaaclab_arena.relations.relations import (
     FaceTo,
-    NextTo,
-    NotNextTo,
     On,
     RandomAroundSolution,
-    RelationBase,
     RotateAroundSolution,
     get_anchor_objects,
+    get_relation,
 )
-from isaaclab_arena.relations.warp_sdf_kernels import has_sdf_sentinel, mesh_sdf
 from isaaclab_arena.utils.bounding_box import AxisAlignedBoundingBox
 from isaaclab_arena.utils.pose import Pose, PosePerEnv
 from isaaclab_arena.utils.random import get_random_rotation
-from isaaclab_arena.utils.yaw import (
-    centers_in_target_frame,
-    rotate_quat_by_yaw,
-    wrap_angle_to_pi,
-    yaw_from_quat_xyzw,
-    yaw_toward_positions,
-)
+from isaaclab_arena.utils.yaw import rotate_quat_by_yaw, wrap_angle_to_pi, yaw_from_quat_xyzw, yaw_toward_positions
 
 if TYPE_CHECKING:
     from isaaclab_arena.assets.object_base import ObjectBase
     from isaaclab_arena.relations.collision_object import CollisionObject
-    from isaaclab_arena.relations.warp_mesh_manager import WarpMeshAndSphereCache
-
-
-RelationT = TypeVar("RelationT", bound=RelationBase)
+    from isaaclab_arena.relations.placement_validation import PlacementValidator
 
 
 @dataclass
@@ -98,7 +82,7 @@ class ObjectPlacer:
     def __init__(self, params: ObjectPlacerParams | None = None):
         self.params = params or ObjectPlacerParams()
         self._solver = RelationSolver(params=self.params.solver_params)
-        self._cpu_mesh_manager: WarpMeshAndSphereCache | None = None
+        self._validators: list[PlacementValidator] = build_validators(self.params.enabled_checks, self.params)
 
     def place(
         self,
@@ -243,11 +227,8 @@ class ObjectPlacer:
         assign_variants_for_envs(objects, num_envs, placement_seed=self.params.placement_seed)
         num_candidates = num_envs * candidates_per_env
         env_bboxes = build_per_env_bounding_boxes(objects, num_envs)
+        env_bboxes = self._enclose_roll_pitch_marker_footprints(objects, env_bboxes)
         unrotated_candidate_bboxes = env_bboxes.get_bounding_boxes_for_solver_candidates(candidates_per_env)
-        # Roll/pitch marker objects carry a fixed out-of-plane rotation the yaw pass cannot
-        # represent; enclose that rotated footprint once so overlap checks match the rotation
-        # _apply_poses writes verbatim. The yaw pass below then leaves these objects untouched.
-        base_candidate_bboxes = self._bake_marker_rotation_footprints(objects, unrotated_candidate_bboxes)
         per_env_bboxes = env_bboxes.get_bounding_boxes_for_all_envs()
 
         initial_positions: list[dict[ObjectBase, tuple[float, float, float]]] = []
@@ -265,7 +246,9 @@ class ObjectPlacer:
             )
 
         # Bake each candidate's yaw into a conservative enclosing bbox for overlap checks.
-        candidate_bboxes = self._rotate_candidate_bboxes(objects, base_candidate_bboxes, orientations_per_candidate)
+        candidate_bboxes = self._rotate_candidate_bboxes(
+            objects, unrotated_candidate_bboxes, orientations_per_candidate
+        )
 
         all_positions = self._solver.solve(
             objects,
@@ -276,19 +259,19 @@ class ObjectPlacer:
             collision_objects=collision_objects,
         )
         self._apply_face_to_orientations(all_positions, orientations_per_candidate)
-        # FaceTo yaw is only known after solving, so rebuild from the base boxes before validation.
-        candidate_bboxes = self._rotate_candidate_bboxes(objects, base_candidate_bboxes, orientations_per_candidate)
+        # FaceTo yaw is only known after solving, so rebuild from unrotated boxes before validation.
+        candidate_bboxes = self._rotate_candidate_bboxes(
+            objects, unrotated_candidate_bboxes, orientations_per_candidate
+        )
         assert self._solver.last_loss_per_env is not None
         all_losses: list[float] = self._solver.last_loss_per_env.cpu().tolist()
-        all_validations = [
-            self._validate_placement(
-                positions,
-                self._get_bounding_boxes_for_candidate_index(candidate_bboxes, candidate_idx),
-                orientations_per_candidate[candidate_idx],
-                collision_objects,
-            )
-            for candidate_idx, positions in enumerate(all_positions)
+        per_candidate_bboxes = [
+            self._get_bounding_boxes_for_candidate_index(candidate_bboxes, candidate_idx)
+            for candidate_idx in range(num_candidates)
         ]
+        all_validations = self._validate_candidates(
+            all_positions, orientations_per_candidate, per_candidate_bboxes, collision_objects
+        )
 
         candidates: list[PlacementCandidate] = []
         for candidate_idx in range(num_candidates):
@@ -414,12 +397,13 @@ class ObjectPlacer:
     ) -> dict[ObjectBase, float]:
         """Sample absolute world Z-yaws for non-anchor objects without FaceTo.
 
-        Marker yaw is included; random_yaw_init adds a sampled delta. Roll/pitch marker objects
-        are omitted here so their requested rotation is applied directly.
+        Marker yaw is included; random_yaw_init adds a sampled delta. Roll/pitch marker objects are
+        omitted so their requested rotation is applied verbatim; their footprint is enclosed by
+        _enclose_roll_pitch_marker_footprints so overlap validation stays sound.
         """
         orientations: dict[ObjectBase, float] = {}
         for obj in objects:
-            marker = self._get_relation(obj, RotateAroundSolution)
+            marker = get_relation(obj, RotateAroundSolution)
             has_roll_pitch = marker is not None and (marker.roll_rad != 0.0 or marker.pitch_rad != 0.0)
             marker_yaw = marker.yaw_rad if marker is not None else 0.0
             if obj in anchor_objects:
@@ -428,7 +412,7 @@ class ObjectPlacer:
                     "Anchors are not repositioned by the placer, so any marker rotation must "
                     "already be baked into the anchor's initial_pose before calling place()."
                 )
-            elif self._get_relation(obj, FaceTo) is None and not has_roll_pitch:
+            elif get_relation(obj, FaceTo) is None and not has_roll_pitch:
                 sampled_yaw = get_random_rotation(generator) if self.params.random_yaw_init else 0.0
                 total_yaw = wrap_angle_to_pi(sampled_yaw + marker_yaw)
                 if total_yaw != 0.0:
@@ -448,7 +432,7 @@ class ObjectPlacer:
         assert len(positions_per_candidate) == len(orientations_per_candidate)
         objects = positions_per_candidate[0]
         for obj in objects:
-            relation = ObjectPlacer._get_relation(obj, FaceTo)
+            relation = get_relation(obj, FaceTo)
             if relation is None:
                 continue
             subject_positions = torch.tensor([positions[obj] for positions in positions_per_candidate])
@@ -457,27 +441,6 @@ class ObjectPlacer:
             for candidate_idx, (yaw, direction_is_defined) in enumerate(zip(yaws, is_defined, strict=True)):
                 if direction_is_defined:
                     orientations_per_candidate[candidate_idx][obj] = yaw.item()
-
-    def _bake_marker_rotation_footprints(
-        self,
-        objects: list[ObjectBase],
-        candidate_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-    ) -> dict[ObjectBase, AxisAlignedBoundingBox]:
-        """Enclose the applied footprint of roll/pitch marker objects.
-
-        Roll/pitch marker objects are excluded from the sampled-yaw path, because the yaw-only bbox
-        pass cannot represent that out-of-plane rotation, so replace each such object's box with the AABB enclosing
-        it after the marker rotation. The subsequent yaw pass leaves these objects untouched.
-
-        Returns a shallow copy with only roll/pitch objects replaced; other boxes are shared.
-        """
-        baked = dict(candidate_bboxes)
-        for obj in objects:
-            marker = self._get_relation(obj, RotateAroundSolution)
-            if marker is None or (marker.roll_rad == 0.0 and marker.pitch_rad == 0.0):
-                continue
-            baked[obj] = candidate_bboxes[obj].enclosing_after_rotation(marker.get_rotation_xyzw())
-        return baked
 
     @staticmethod
     def _rotate_candidate_bboxes(
@@ -502,6 +465,23 @@ class ObjectPlacer:
                 bbox = bbox.rotated_around_z(yaw_tensor)
             rotated[obj] = bbox
         return rotated
+
+    @staticmethod
+    def _enclose_roll_pitch_marker_footprints(
+        objects: list[ObjectBase], env_bboxes: PerEnvBoundingBoxes
+    ) -> PerEnvBoundingBoxes:
+        """Refit roll/pitch RotateAroundSolution objects' bboxes to enclose their marker-rotated box.
+
+        These objects keep their requested rotation without a sampled yaw, but a Z-only footprint
+        can't represent a tilted box, so overlap validation must use the box enclosing the full
+        marker rotation. Other objects are returned unchanged.
+        """
+        rotated_bboxes = dict(env_bboxes.object_bboxes)
+        for obj in objects:
+            marker = get_relation(obj, RotateAroundSolution)
+            if marker is not None and (marker.roll_rad != 0.0 or marker.pitch_rad != 0.0):
+                rotated_bboxes[obj] = rotated_bboxes[obj].rotated_by_quat(marker.get_rotation_xyzw())
+        return PerEnvBoundingBoxes(object_bboxes=rotated_bboxes, num_envs=env_bboxes.num_envs)
 
     @staticmethod
     def _get_bounding_boxes_for_candidate_index(
@@ -606,568 +586,38 @@ class ObjectPlacer:
             return float((parent_min + parent_max) / 2.0)
         return float(low + (high - low) * torch.rand(1, generator=generator).item())
 
-    def _validate_on_relations(
+    def _validate_candidates(
         self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        env_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-    ) -> bool:
-        """Validate each On relation; keep in sync with OnLossStrategy in relation_loss_strategies.py.
+        positions: list[dict[ObjectBase, tuple[float, float, float]]],
+        orientations: list[dict[ObjectBase, float]],
+        bboxes: list[dict[ObjectBase, AxisAlignedBoundingBox]],
+        collision_objects: list[CollisionObject],
+    ) -> list[PlacementValidationResults]:
+        """Run every enabled validator over all candidates and collect per-candidate results.
 
-        1. X: child's footprint within parent's X extent, inset by the relation's edge_margin_m.
-        2. Y: child's footprint within parent's Y extent, inset by the relation's edge_margin_m.
-        3. Z: child_bottom in (parent_top, parent_top+clearance_m], within on_relation_z_tolerance_m.
+        Each validator reports one verdict per candidate; the verdicts are transposed into one
+        PlacementValidationResults per candidate, gated by the configured required_checks.
 
         Args:
-            positions: Solved positions for each object.
-            env_bboxes: Per-object bboxes for the current env, each with shape (1, 3).
+            positions: Solved (x, y, z) per object, one dict per candidate.
+            orientations: Absolute world Z-yaw per object, one dict per candidate (may be empty).
+            bboxes: Per-object bboxes for each candidate's env, each (1, 3).
+            collision_objects: Fixed background obstacles shared across candidates.
         """
-        for obj in positions:
-            for rel in obj.get_relations():
-                if not isinstance(rel, On):
-                    continue
-                parent = rel.parent
-                if parent not in positions:
-                    continue
-                child_bbox = env_bboxes[obj]
-                parent_bbox = env_bboxes[parent]
-                child_world = child_bbox.translated(positions[obj])
-                parent_world = parent_bbox.translated(positions[parent])
-                parent_size = parent_world.max_point - parent_world.min_point
-                child_size = child_world.max_point - child_world.min_point
-
-                m = rel.edge_margin_m
-                # 1) Checking that with the specified margin, the parent is wide enough to place the child on top
-                if m > 0.0:
-                    freespace = parent_size - child_size
-                    # A margin too large for the surface inverts the inset band so containment can never pass.
-                    if torch.any(freespace[0, :2] < 2 * m):
-                        # The maximum feasible margin is the minimum of the freespace on the xy axes.
-                        max_feasible_margin = max(0.0, min(freespace[0, :2]) / 2.0)
-                        # When parent < child, freespace[0, :2] is negative and max_feasible_margin is 0.0.
-                        if max_feasible_margin > 0.0:
-                            if self.params.verbose:
-                                print(
-                                    f"On relation: edge_margin_m={m} m is too large for parent '{parent.name}'. Max"
-                                    f" feasible margin here is {max_feasible_margin:.3f} m. Use a smaller"
-                                    " edge_margin_m."
-                                )
-                            return False
-                # 2) Checking that the child lies within the parent's xy
-                if (
-                    child_world.min_point[0, 0] < parent_world.min_point[0, 0] + m
-                    or child_world.max_point[0, 0] > parent_world.max_point[0, 0] - m
-                    or child_world.min_point[0, 1] < parent_world.min_point[0, 1] + m
-                    or child_world.max_point[0, 1] > parent_world.max_point[0, 1] - m
-                ):
-                    if self.params.verbose:
-                        print(f"On relation: '{obj.name}' XY outside parent (retrying)")
-                    return False
-                # 3) Checking that the child lies within an acceptable z-range.
-                parent_local_top_z: float = parent_bbox.max_point[0, 2].item()
-                child_local_bottom_z: float = child_bbox.min_point[0, 2].item()
-                parent_top_z = parent_local_top_z + positions[parent][2]
-                clearance_m = rel.clearance_m
-                child_bottom_z = child_local_bottom_z + positions[obj][2]
-                eps_z = self.params.on_relation_z_tolerance_m
-                if child_bottom_z <= parent_top_z - eps_z or child_bottom_z > parent_top_z + clearance_m + eps_z:
-                    if self.params.verbose:
-                        print(f"  On relation: '{obj.name}' Z outside band (retrying)")
-                    return False
-        return True
-
-    @staticmethod
-    def _collect_skip_pairs(
-        positions: dict[ObjectBase, tuple[float, float, float]],
-    ) -> tuple[set[tuple], set[int]]:
-        """Build On-pair skip set and anchor ID set from positioned objects.
-
-        Returns:
-            Tuple of (on_pairs, anchor_ids) where on_pairs contains (id(a), id(b))
-            tuples for On-linked objects, and anchor_ids contains id() of anchors.
-        """
-        on_pairs: set[tuple] = set()
-        anchor_ids: set[int] = set()
-        for obj in positions:
-            for rel in obj.get_relations():
-                if isinstance(rel, On) and rel.parent in positions:
-                    on_pairs.add((id(obj), id(rel.parent)))
-                    on_pairs.add((id(rel.parent), id(obj)))
-            if obj.is_anchor:
-                anchor_ids.add(id(obj))
-        return on_pairs, anchor_ids
-
-    def _non_skip_pairs(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        skip_mesh_pairs: bool = False,
-    ) -> Iterator[tuple[ObjectBase, ObjectBase]]:
-        """Yield non-relation object pairs, optionally skipping pairs handled by mesh collision."""
-        on_pairs, anchor_ids = self._collect_skip_pairs(positions)
-        mesh_manager = self._get_cpu_mesh_manager() if skip_mesh_pairs else None
-        default_collision_mode = self.params.solver_params.collision_mode
-        objects = list(positions.keys())
-        for i in range(len(objects)):
-            for j in range(i + 1, len(objects)):
-                a, b = objects[i], objects[j]
-                if id(a) in anchor_ids and id(b) in anchor_ids:
-                    continue
-                if (id(a), id(b)) in on_pairs:
-                    continue
-                if mesh_manager is not None and (
-                    (
-                        object_uses_mesh_collision(a, default_collision_mode)
-                        and mesh_manager.get_collision_mesh(a) is not None
-                    )
-                    or (
-                        object_uses_mesh_collision(b, default_collision_mode)
-                        and mesh_manager.get_collision_mesh(b) is not None
-                    )
-                ):
-                    continue
-                yield a, b
-
-    def _validate_no_overlap(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        env_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-        collision_objects: list[CollisionObject] | None = None,
-        skip_mesh_pairs: bool = False,
-    ) -> bool:
-        """AABB overlap check on pre-rotated env_bboxes. Skips On-pairs and anchor-anchor pairs."""
-        clearance_m = self.params.solver_params.clearance_m
-        margin = max(0.0, clearance_m - 1e-6)
-        collision_objects = collision_objects or []
-        _, anchor_ids = self._collect_skip_pairs(positions)
-
-        for a, b in self._non_skip_pairs(positions, skip_mesh_pairs=skip_mesh_pairs):
-            if self._pair_aabb_overlaps(env_bboxes[a], env_bboxes[b], positions[a], positions[b], 0.0, 0.0, margin):
-                if self.params.verbose:
-                    print(f"  Overlap between '{a.name}' and '{b.name}'")
-                return False
-
-        # Placed (non-anchor) objects must also clear the fixed background obstacles.
-        # Anchors are fixed scene geometry too, so anchor-vs-background overlap is not gated.
-        background_worlds = [(bg, bg.get_world_bounding_box()) for bg in collision_objects]
-        mesh_manager = self._get_cpu_mesh_manager() if skip_mesh_pairs else None
-        default_collision_mode = self.params.solver_params.collision_mode
-        for obj in positions:
-            if id(obj) in anchor_ids:
-                continue
-            obj_world = env_bboxes[obj].translated(positions[obj])
-            for background, background_world in background_worlds:
-                if (
-                    mesh_manager is not None
-                    and object_uses_mesh_collision(background, default_collision_mode)
-                    and mesh_manager.get_collision_mesh(background) is not None
-                ):
-                    continue
-                if obj_world.overlaps(background_world, margin=margin).item():
-                    if self.params.verbose:
-                        print(f"  Overlap between '{obj.name}' and background '{background.name}'")
-                    return False
-        return True
-
-    def _validate_next_to_relations(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        env_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-    ) -> bool:
-        """Validate each NextTo relation: child on the requested side, facing edge within the
-        relation's tolerance_m of distance_m from the parent edge. Shares next_to_violations with
-        NextToLossStrategy; cross_position_ratio is a soft preference and is not gated.
-
-        Args:
-            positions: Solved positions for each object.
-            env_bboxes: Per-object bboxes for the current env, each with shape (1, 3).
-        """
-        for obj in positions:
-            for rel in obj.get_relations():
-                if not isinstance(rel, NextTo):
-                    continue
-                parent = rel.parent
-                if parent not in positions:
-                    continue
-                cfg = SIDE_CONFIGS[rel.side]
-                child_bbox = env_bboxes[obj]
-                child_pos = child_bbox.min_point.new_tensor([positions[obj]])
-                parent_world = env_bboxes[parent].translated(positions[parent])
-                half_plane, distance = next_to_violations(cfg, child_pos, child_bbox, parent_world, rel.distance_m)
-
-                if half_plane.item() > rel.tolerance_m or distance.item() > rel.tolerance_m:
-                    if self.params.verbose:
-                        print(
-                            f"NextTo: '{obj.name}' next_to({parent.name}) violated"
-                            f" (side={half_plane.item():.4f}, distance={distance.item():.4f} m;"
-                            f" tolerance_m={rel.tolerance_m})"
-                        )
-                    return False
-        return True
-
-    def _validate_not_next_to_relations(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        env_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-    ) -> bool:
-        """Validate each NotNextTo relation: child has cleared the keep-out zone beside the parent
-        (within the relation's tolerance_m) via either route — back over the edge or past the
-        footprint end. Shares not_next_to_violations with NotNextToLossStrategy, using its margin_m.
-
-        Args:
-            positions: Solved positions for each object.
-            env_bboxes: Per-object bboxes for the current env, each with shape (1, 3).
-        """
-        for obj in positions:
-            for rel in obj.get_relations():
-                if not isinstance(rel, NotNextTo):
-                    continue
-                parent = rel.parent
-                if parent not in positions:
-                    continue
-                cfg = SIDE_CONFIGS[rel.side]
-                margin_m = self._not_next_to_margin(rel)
-                child_bbox = env_bboxes[obj]
-                child_pos = child_bbox.min_point.new_tensor([positions[obj]])
-                parent_world = env_bboxes[parent].translated(positions[parent])
-                remaining_side, remaining_cross = not_next_to_violations(
-                    cfg, child_pos, child_bbox, parent_world, margin_m
-                )
-
-                if min(remaining_side.item(), remaining_cross.item()) > rel.tolerance_m:
-                    if self.params.verbose:
-                        print(
-                            f"NotNextTo: '{obj.name}' not_next_to({parent.name}) violated"
-                            f" (remaining_side={remaining_side.item():.4f},"
-                            f" remaining_cross={remaining_cross.item():.4f} m;"
-                            f" margin_m={margin_m}, tolerance_m={rel.tolerance_m})"
-                        )
-                    return False
-        return True
-
-    def _not_next_to_margin(self, relation: NotNextTo) -> float:
-        """Keep-out margin_m from the registered NotNextTo loss strategy (stays in sync with the solver)."""
-        strategy = cast(NotNextToLossStrategy, self._solver.params.strategies[type(relation)])
-        return strategy.margin_m
-
-    def _get_cpu_mesh_manager(self) -> WarpMeshAndSphereCache:
-        """Return the CPU-device mesh manager, creating it on first call."""
-        if self._cpu_mesh_manager is None:
-            from isaaclab_arena.relations.warp_mesh_manager import WarpMeshAndSphereCache
-
-            self._cpu_mesh_manager = WarpMeshAndSphereCache(
-                num_spheres=self.params.solver_params.num_spheres,
-                device="cpu",
+        verdicts_by_check = {
+            validator.check: validator.validate_batch(positions, orientations, bboxes, collision_objects)
+            for validator in self._validators
+        }
+        # required_checks=None means "every enabled check is required"; an empty set on the results
+        # already carries that meaning (see PlacementValidationResults._required).
+        required = self.params.required_checks
+        return [
+            PlacementValidationResults(
+                validation_results={check: verdicts[candidate_idx] for check, verdicts in verdicts_by_check.items()},
+                required_checks=set(required) if required is not None else set(),
             )
-        return self._cpu_mesh_manager
-
-    def _validate_no_overlap_mesh(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        env_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-        orientations: dict[ObjectBase, float] | None = None,
-        collision_objects: list[CollisionObject] | None = None,
-    ) -> bool:
-        """Sphere-to-SDF overlap check; both-meshless pairs fall back to AABB validation."""
-        clearance_m = self.params.solver_params.clearance_m
-        tolerance = max(0.0, clearance_m - 1e-6)
-        mesh_manager = self._get_cpu_mesh_manager()
-        mesh_manager.reset_sentinel_warning()
-        warned_no_mesh: set[str] = set()
-        collision_objects = collision_objects or []
-        default_collision_mode = self.params.solver_params.collision_mode
-
-        for a, b in self._non_skip_pairs(positions):
-            a_uses_mesh = object_uses_mesh_collision(a, default_collision_mode)
-            b_uses_mesh = object_uses_mesh_collision(b, default_collision_mode)
-            a_mesh = mesh_manager.get_collision_mesh(a) if a_uses_mesh else None
-            b_mesh = mesh_manager.get_collision_mesh(b) if b_uses_mesh else None
-            if a_mesh is None and b_mesh is None:
-                for obj, uses_mesh, mesh in [(a, a_uses_mesh, a_mesh), (b, b_uses_mesh, b_mesh)]:
-                    if uses_mesh and mesh is None and obj.name not in warned_no_mesh:
-                        warned_no_mesh.add(obj.name)
-                        print(
-                            f"  [NoCollision] MESH mode: '{obj.name}' has no collision mesh,"
-                            " falling back to AABB validation for this pair"
-                        )
-                continue
-
-            a_pos = torch.tensor(positions[a], dtype=torch.float32)
-            b_pos = torch.tensor(positions[b], dtype=torch.float32)
-
-            if b_mesh is not None and self._spheres_penetrate_mesh(
-                a,
-                self._collision_mesh_or_aabb_proxy(a_mesh, env_bboxes[a]),
-                a if a_mesh is not None else None,
-                a_mesh is not None,
-                a.is_anchor,
-                a_pos,
-                b,
-                b_mesh,
-                b_pos,
-                b.is_anchor,
-                mesh_manager,
-                tolerance,
-                orientations,
-            ):
-                return False
-            if a_mesh is not None and self._spheres_penetrate_mesh(
-                b,
-                self._collision_mesh_or_aabb_proxy(b_mesh, env_bboxes[b]),
-                b if b_mesh is not None else None,
-                b_mesh is not None,
-                b.is_anchor,
-                b_pos,
-                a,
-                a_mesh,
-                a_pos,
-                a.is_anchor,
-                mesh_manager,
-                tolerance,
-                orientations,
-            ):
-                return False
-
-        for source in positions:
-            if source.is_anchor:
-                continue
-            source_mesh = (
-                mesh_manager.get_collision_mesh(source)
-                if object_uses_mesh_collision(source, default_collision_mode)
-                else None
-            )
-            source_pos = torch.tensor(positions[source], dtype=torch.float32)
-            for background in collision_objects:
-                target_mesh = (
-                    mesh_manager.get_collision_mesh(background)
-                    if object_uses_mesh_collision(background, default_collision_mode)
-                    else None
-                )
-                if target_mesh is None:
-                    continue
-                target_pose = background.get_initial_pose()
-                assert isinstance(
-                    target_pose, Pose
-                ), f"Background collision object '{background.name}' must have a fixed Pose in MESH mode."
-                target_pos = torch.tensor(target_pose.position_xyz, dtype=torch.float32)
-                if self._spheres_penetrate_mesh(
-                    source,
-                    self._collision_mesh_or_aabb_proxy(source_mesh, env_bboxes[source]),
-                    source if source_mesh is not None else None,
-                    source_mesh is not None,
-                    source.is_anchor,
-                    source_pos,
-                    background,
-                    target_mesh,
-                    target_pos,
-                    True,
-                    mesh_manager,
-                    tolerance,
-                    orientations,
-                ):
-                    return False
-
-        return True
-
-    @staticmethod
-    def _collision_mesh_or_aabb_proxy(
-        mesh: trimesh.Trimesh | None,
-        bbox: AxisAlignedBoundingBox,
-    ) -> trimesh.Trimesh:
-        """Return an object's collision mesh, or a box mesh matching the candidate AABB."""
-        if mesh is not None:
-            return mesh
-        box_mesh = trimesh.creation.box(extents=bbox.size[0].detach().cpu().numpy())
-        box_mesh.apply_translation(bbox.center[0].detach().cpu().numpy())
-        return box_mesh
-
-    def _spheres_penetrate_mesh(
-        self,
-        source: ObjectBase,
-        source_mesh: trimesh.Trimesh,
-        source_sphere_cache_obj: ObjectBase | None,
-        source_applies_yaw: bool,
-        source_uses_pose_yaw: bool,
-        source_pos: torch.Tensor,
-        target: ObjectBase | CollisionObject,
-        target_mesh: trimesh.Trimesh,
-        target_pos: torch.Tensor,
-        target_uses_pose_yaw: bool,
-        mesh_manager: WarpMeshAndSphereCache,
-        tolerance: float,
-        orientations: dict[ObjectBase, float] | None,
-    ) -> bool:
-        """True if source's spheres penetrate target's mesh or if BVH returns no-face sentinel.
-
-        source_applies_yaw describes whether sphere centers need sampled-yaw rotation.
-        *_uses_pose_yaw controls whether fixed anchors/passive obstacles contribute pose yaw.
-        """
-        spheres = mesh_manager.get_query_spheres(source_mesh, obj=source_sphere_cache_obj)
-        warp_mesh = mesh_manager.get_warp_mesh(target_mesh, obj=target)
-        centers = self._centers_in_target_frame(
-            spheres[:, :3],
-            source,
-            target,
-            source_pos,
-            target_pos,
-            orientations,
-            source_applies_yaw=source_applies_yaw,
-            source_uses_pose_yaw=source_uses_pose_yaw,
-            target_uses_pose_yaw=target_uses_pose_yaw,
-        )
-        sdf = mesh_sdf(centers, warp_mesh)
-        mesh_manager.warn_sdf_sentinel(sdf)
-        if has_sdf_sentinel(sdf):
-            return True
-        if (sdf < spheres[:, 3] + tolerance).any():
-            if self.params.verbose:
-                print(f"  Mesh overlap between '{source.name}' and '{target.name}'")
-            return True
-        return False
-
-    @staticmethod
-    def _effective_yaw(
-        obj: ObjectBase | CollisionObject,
-        orientations: dict[ObjectBase, float] | None,
-        use_pose_yaw: bool,
-    ) -> float:
-        """Resolve effective Z-yaw from sampled orientations or, when allowed, fixed initial pose."""
-        if orientations is not None and obj in orientations:
-            return orientations[cast("ObjectBase", obj)]
-        if not use_pose_yaw:
-            return 0.0
-        pose = obj.get_initial_pose()
-        if not isinstance(pose, Pose):
-            return 0.0
-        return yaw_from_quat_xyzw(pose.rotation_xyzw)
-
-    @staticmethod
-    def _pair_aabb_overlaps(
-        a_bbox: AxisAlignedBoundingBox,
-        b_bbox: AxisAlignedBoundingBox,
-        pos_a: tuple[float, float, float],
-        pos_b: tuple[float, float, float],
-        yaw_a: float,
-        yaw_b: float,
-        margin: float,
-    ) -> bool:
-        """True if two yaw-rotated, world-translated AABBs overlap (with margin)."""
-        if yaw_a != 0.0:
-            a_bbox = a_bbox.rotated_around_z(yaw_a)
-        if yaw_b != 0.0:
-            b_bbox = b_bbox.rotated_around_z(yaw_b)
-        a_world = a_bbox.translated(pos_a)
-        b_world = b_bbox.translated(pos_b)
-        return a_world.overlaps(b_world, margin=margin).item()
-
-    @staticmethod
-    def _centers_in_target_frame(
-        centers_local: torch.Tensor,
-        source_obj: ObjectBase,
-        target_obj: ObjectBase | CollisionObject,
-        source_pos: torch.Tensor,
-        target_pos: torch.Tensor,
-        orientations: dict[ObjectBase, float] | None,
-        source_applies_yaw: bool = True,
-        source_uses_pose_yaw: bool = True,
-        target_uses_pose_yaw: bool = True,
-    ) -> torch.Tensor:
-        """Transform source sphere centers into the target's local frame (Z-yaw only)."""
-        src_yaw = (
-            ObjectPlacer._effective_yaw(source_obj, orientations, source_uses_pose_yaw) if source_applies_yaw else 0.0
-        )
-        tgt_yaw = ObjectPlacer._effective_yaw(target_obj, orientations, target_uses_pose_yaw)
-        return centers_in_target_frame(centers_local, src_yaw, tgt_yaw, source_pos - target_pos)
-
-    def _validate_placement(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        env_bboxes: dict[ObjectBase, AxisAlignedBoundingBox],
-        orientations: dict[ObjectBase, float] | None = None,
-        collision_objects: list[CollisionObject] | None = None,
-    ) -> PlacementValidationResults:
-        """Validate overlap and all supported placement relations.
-
-        Args:
-            positions: Dictionary mapping objects to their solved (x, y, z) positions.
-            env_bboxes: Per-object bboxes for the current env, each with shape (1, 3).
-            orientations: Optional per-object yaw (radians about Z).
-            collision_objects: Fixed background obstacles to test placed objects against.
-
-        Returns:
-            PlacementValidationResults with the overlap and relation checks.
-        """
-        use_mesh = self._should_validate_mesh(positions, collision_objects)
-        no_overlap = self._validate_no_overlap(
-            positions,
-            env_bboxes,
-            collision_objects=collision_objects,
-            skip_mesh_pairs=use_mesh,
-        )
-        if no_overlap and use_mesh:
-            no_overlap = self._validate_no_overlap_mesh(positions, env_bboxes, orientations, collision_objects)
-        on_relation = self._validate_on_relations(positions, env_bboxes)
-        next_to = self._validate_next_to_relations(positions, env_bboxes)
-        not_next_to = self._validate_not_next_to_relations(positions, env_bboxes)
-        face_to = self._validate_face_to_relations(positions, orientations)
-
-        validation_results = cast(
-            dict[PlacementCheck, bool],
-            {
-                PlacementCheck.NO_OVERLAP: no_overlap,
-                PlacementCheck.ON_RELATION: on_relation,
-                PlacementCheck.NEXT_TO: next_to,
-                PlacementCheck.NOT_NEXT_TO: not_next_to,
-                PlacementCheck.FACE_TO: face_to,
-            },
-        )
-        required_checks = cast(
-            set[PlacementCheck],
-            {
-                PlacementCheck.NO_OVERLAP,
-                PlacementCheck.ON_RELATION,
-                PlacementCheck.NEXT_TO,
-                PlacementCheck.NOT_NEXT_TO,
-                PlacementCheck.FACE_TO,
-            },
-        )
-        return PlacementValidationResults(
-            validation_results=validation_results,
-            required_checks=required_checks,
-        )
-
-    def _validate_face_to_relations(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        orientations: dict[ObjectBase, float] | None,
-    ) -> bool:
-        """Validate that every FaceTo subject has a defined direction and computed yaw."""
-        for obj in positions:
-            face_to = self._get_relation(obj, FaceTo)
-            if face_to is None:
-                continue
-            subject_position = torch.tensor([positions[obj]])
-            target_position = torch.tensor([positions[face_to.parent]])
-            _, direction_is_defined = yaw_toward_positions(subject_position, target_position)
-            if not direction_is_defined.item():
-                if self.params.verbose:
-                    print(f"  FaceTo: '{obj.name}' is too close to its target in XY")
-                return False
-            if orientations is None or obj not in orientations:
-                if self.params.verbose:
-                    print(f"  FaceTo: '{obj.name}' has no computed facing yaw")
-                return False
-        return True
-
-    def _should_validate_mesh(
-        self,
-        positions: dict[ObjectBase, tuple[float, float, float]],
-        collision_objects: list[CollisionObject] | None,
-    ) -> bool:
-        """Return True when any object in this validation uses mesh collision."""
-        default_collision_mode = self.params.solver_params.collision_mode
-        if default_collision_mode == CollisionMode.MESH:
-            return True
-        objects = [*positions.keys(), *(collision_objects or [])]
-        return any(get_object_collision_mode(obj, default_collision_mode) == CollisionMode.MESH for obj in objects)
+            for candidate_idx in range(len(positions))
+        ]
 
     def _apply_poses(
         self,
@@ -1185,7 +635,7 @@ class ObjectPlacer:
             if obj in anchor_objects:
                 continue
 
-            rotate_marker = self._get_relation(obj, RotateAroundSolution)
+            rotate_marker = get_relation(obj, RotateAroundSolution)
             marker_rotation = rotate_marker.get_rotation_xyzw() if rotate_marker else (0.0, 0.0, 0.0, 1.0)
             marker_yaw = yaw_from_quat_xyzw(marker_rotation)
 
@@ -1196,7 +646,7 @@ class ObjectPlacer:
             if num_envs == 1:
                 pos = positions_per_env[0][obj]
                 rotation_xyzw = rotate_quat_by_yaw(marker_rotation, _yaw_delta(0))
-                random_marker = self._get_relation(obj, RandomAroundSolution)
+                random_marker = get_relation(obj, RandomAroundSolution)
                 if random_marker is not None:
                     obj.set_initial_pose(random_marker.to_pose_range_centered_at(pos, rotation_xyzw=rotation_xyzw))
                 else:
@@ -1210,10 +660,6 @@ class ObjectPlacer:
                     for env_idx in range(num_envs)
                 ]
                 obj.set_initial_pose(PosePerEnv(poses=poses))
-
-    @staticmethod
-    def _get_relation(obj: ObjectBase, relation_type: type[RelationT]) -> RelationT | None:
-        return next((relation for relation in obj.get_relations() if isinstance(relation, relation_type)), None)
 
     @property
     def last_loss_history(self) -> list[float]:
