+++ isaaclab_arena/tests/test_face_to.py
@@ -20,6 +20,7 @@ from isaaclab_arena.relations.collision_mode import CollisionMode
 from isaaclab_arena.relations.object_placer import ObjectPlacer
 from isaaclab_arena.relations.object_placer_params import ObjectPlacerParams
 from isaaclab_arena.relations.placement_validation import PlacementCheck
+from isaaclab_arena.relations.placement_validators import NoOverlapValidator
 from isaaclab_arena.relations.relation_solver import RelationSolver
 from isaaclab_arena.relations.relation_solver_params import RelationSolverParams
 from isaaclab_arena.relations.relations import AtPosition, FaceTo, IsAnchor, RandomAroundSolution, RotateAroundSolution
@@ -137,11 +138,12 @@ def test_coincident_face_to_fails_candidate_validation():
     orientations = [{}]
 
     ObjectPlacer._apply_face_to_orientations([positions], orientations)
-    validation = ObjectPlacer()._validate_placement(
-        positions,
-        {obj: obj.get_bounding_box() for obj in positions},
-        {subject: 0.0},
-    )
+    validation = ObjectPlacer()._validate_candidates(
+        [positions],
+        [{subject: 0.0}],
+        [{obj: obj.get_bounding_box() for obj in positions}],
+        [],
+    )[0]
 
     assert orientations == [{}]
     assert validation.validation_results[PlacementCheck.FACE_TO] is False
@@ -159,10 +161,12 @@ def test_face_to_rebuilds_rotated_footprint_before_validation():
     orientations = [{}]
     placer = ObjectPlacer()
 
-    assert placer._validate_placement(positions, unrotated).validation_results[PlacementCheck.NO_OVERLAP]
+    assert placer._validate_candidates([positions], [{}], [unrotated], [])[0].validation_results[
+        PlacementCheck.NO_OVERLAP
+    ]
     ObjectPlacer._apply_face_to_orientations([positions], orientations)
     rotated = ObjectPlacer._rotate_candidate_bboxes(objects, unrotated, orientations)
-    validation = placer._validate_placement(positions, rotated, orientations[0])
+    validation = placer._validate_candidates([positions], [orientations[0]], [rotated], [])[0]
 
     assert orientations[0][subject] == pytest.approx(math.pi / 2)
     assert validation.validation_results[PlacementCheck.NO_OVERLAP] is False
@@ -340,7 +344,8 @@ def test_mesh_validation_receives_final_face_to_yaw(monkeypatch):
         received.update(candidate_orientations or {})
         return True
 
-    monkeypatch.setattr(placer, "_validate_no_overlap_mesh", _capture_orientations)
+    no_overlap_validator = next(v for v in placer._validators if isinstance(v, NoOverlapValidator))
+    monkeypatch.setattr(no_overlap_validator, "_validate_no_overlap_mesh", _capture_orientations)
     (result,) = placer.place([pair.target, pair.subject])
 
     assert received[pair.subject] == pytest.approx(math.pi / 2)
