+++ isaaclab_arena/tests/test_validate_placement.py
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Tests for ObjectPlacer placement validation (_validate_placement, _validate_no_overlap, _validate_on_relations)."""
+"""Tests for build-time placement validators (overlap, On, NextTo, NotNextTo) and their aggregation."""
 
 import math
 import torch
@@ -12,6 +12,7 @@ from isaaclab_arena.assets.dummy_object import DummyObject
 from isaaclab_arena.relations.object_placer import ObjectPlacer
 from isaaclab_arena.relations.object_placer_params import ObjectPlacerParams
 from isaaclab_arena.relations.placement_validation import PlacementCheck
+from isaaclab_arena.relations.placement_validators import NextToValidator, NotNextToValidator, OnRelationValidator
 from isaaclab_arena.relations.relations import NextTo, NotNextTo, On, RotateAroundSolution, Side
 from isaaclab_arena.utils.bounding_box import AxisAlignedBoundingBox
 
@@ -42,6 +43,11 @@ def _env_bboxes(positions: dict[DummyObject, tuple[float, float, float]]):
     return {obj: obj.get_bounding_box() for obj in positions}
 
 
+def _validate_one(placer: ObjectPlacer, positions, env_bboxes, orientations=None):
+    """Run every enabled validator over a single candidate and return its aggregated results."""
+    return placer._validate_candidates([positions], [orientations or {}], [env_bboxes], [])[0]
+
+
 def _stack_rows(bbox: AxisAlignedBoundingBox, n: int) -> AxisAlignedBoundingBox:
     """Repeat a single-env bbox into n stacked rows (one per candidate)."""
     return AxisAlignedBoundingBox(min_point=bbox.min_point.repeat(n, 1), max_point=bbox.max_point.repeat(n, 1))
@@ -53,9 +59,7 @@ def test_no_overlap_returns_true():
     a = _make_box("a")
     b = _make_box("b")
     positions = {a: (0.0, 0.0, 0.0), b: (1.0, 0.0, 0.0)}
-    assert (
-        placer._validate_placement(positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is True
-    )
+    assert _validate_one(placer, positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is True
 
 
 def test_overlapping_returns_false():
@@ -64,9 +68,7 @@ def test_overlapping_returns_false():
     a = _make_box("a")
     b = _make_box("b")
     positions = {a: (0.0, 0.0, 0.0), b: (0.0, 0.0, 0.0)}
-    assert (
-        placer._validate_placement(positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is False
-    )
+    assert _validate_one(placer, positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is False
 
 
 def test_partial_overlap_returns_false():
@@ -75,9 +77,7 @@ def test_partial_overlap_returns_false():
     a = _make_box("a", size=0.2)
     b = _make_box("b", size=0.2)
     positions = {a: (0.0, 0.0, 0.0), b: (0.1, 0.1, 0.0)}
-    assert (
-        placer._validate_placement(positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is False
-    )
+    assert _validate_one(placer, positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is False
 
 
 def test_separated_in_z_passes():
@@ -86,9 +86,7 @@ def test_separated_in_z_passes():
     a = _make_box("a")
     b = _make_box("b")
     positions = {a: (0.0, 0.0, 0.0), b: (0.0, 0.0, 5.0)}
-    assert (
-        placer._validate_placement(positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is True
-    )
+    assert _validate_one(placer, positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is True
 
 
 def test_object_on_surface_no_overlap():
@@ -98,9 +96,7 @@ def test_object_on_surface_no_overlap():
     box = _make_box("box", size=0.2)
     # Desk top at z=0.05; box at z=0.16 → box occupies z=[0.06, 0.26], clear of desk
     positions = {desk: (0.0, 0.0, 0.0), box: (0.0, 0.0, 0.16)}
-    assert (
-        placer._validate_placement(positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is True
-    )
+    assert _validate_one(placer, positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is True
 
 
 def test_colocated_siblings_overlap_rejected():
@@ -110,9 +106,7 @@ def test_colocated_siblings_overlap_rejected():
     a = _make_box("a", size=0.2)
     b = _make_box("b", size=0.2)
     positions = {desk: (0.0, 0.0, 0.0), a: (0.0, 0.0, 0.15), b: (0.0, 0.0, 0.15)}
-    assert (
-        placer._validate_placement(positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is False
-    )
+    assert _validate_one(placer, positions, _env_bboxes(positions)).do_all_required_validation_checks_pass() is False
 
 
 def test_rotation_aware_overlap_uses_yaw():
@@ -122,9 +116,9 @@ def test_rotation_aware_overlap_uses_yaw():
     b = _make_box("b", size=0.1)
     positions = {a: (0.0, 0.0, 0.0), b: (0.0, 0.2, 0.0)}
     axis_aligned = {a: a.get_bounding_box(), b: b.get_bounding_box()}
-    assert placer._validate_placement(positions, axis_aligned).do_all_required_validation_checks_pass() is True
+    assert _validate_one(placer, positions, axis_aligned).do_all_required_validation_checks_pass() is True
     rotated = {a: a.get_bounding_box().rotated_around_z(math.pi / 2), b: b.get_bounding_box()}
-    assert placer._validate_placement(positions, rotated).do_all_required_validation_checks_pass() is False
+    assert _validate_one(placer, positions, rotated).do_all_required_validation_checks_pass() is False
 
 
 def test_candidate_bbox_aligns_with_candidate_yaw():
@@ -140,8 +134,8 @@ def test_candidate_bbox_aligns_with_candidate_yaw():
 
     # Mirrors _place_ranked: each candidate validates against its own bbox row.
     validations = [
-        placer._validate_placement(
-            positions, ObjectPlacer._get_bounding_boxes_for_candidate_index(rotated, idx)
+        _validate_one(
+            placer, positions, ObjectPlacer._get_bounding_boxes_for_candidate_index(rotated, idx)
         ).do_all_required_validation_checks_pass()
         for idx in range(2)
     ]
@@ -212,9 +206,9 @@ def test_on_relation_containment_uses_rotated_bbox():
     positions = {desk: (0.0, 0.0, 0.0), child: (0.0, 0.44, 0.105)}
 
     axis_aligned = {desk: desk.get_bounding_box(), child: child.get_bounding_box()}
-    assert placer._validate_on_relations(positions, axis_aligned) is True
+    assert OnRelationValidator(placer.params)._validate(positions, axis_aligned) is True
     rotated = {desk: desk.get_bounding_box(), child: child.get_bounding_box().rotated_around_z(math.pi / 2)}
-    assert placer._validate_on_relations(positions, rotated) is False
+    assert OnRelationValidator(placer.params)._validate(positions, rotated) is False
 
 
 def test_on_relation_check_no_relation_returns_true():
@@ -223,7 +217,7 @@ def test_on_relation_check_no_relation_returns_true():
     a = _make_box("a")
     b = _make_box("b")
     positions = {a: (0.0, 0.0, 0.0), b: (1.0, 0.0, 0.0)}
-    assert placer._validate_on_relations(positions, _env_bboxes(positions)) is True
+    assert OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_on_relation_check_child_inside_xy_z_in_band_passes():
@@ -235,7 +229,7 @@ def test_on_relation_check_child_inside_xy_z_in_band_passes():
     box.add_relation(On(desk))  # clearance_m=0.01; desk top 0.05
     # Child bottom 0.06 (at upper bound); box half-height 0.1 → center z = 0.16.
     positions = {desk: (0.0, 0.0, 0.0), box: (0.0, 0.0, 0.16)}
-    assert placer._validate_on_relations(positions, _env_bboxes(positions)) is True
+    assert OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_validate_on_relations_child_z_above_clearance_fails():
@@ -247,7 +241,7 @@ def test_validate_on_relations_child_z_above_clearance_fails():
     box.add_relation(On(desk))  # clearance_m=0.01; desk top 0.05
     # Child bottom 1.0 is above band.
     positions = {desk: (0.0, 0.0, 0.0), box: (0.0, 0.0, 1.1)}
-    assert placer._validate_on_relations(positions, _env_bboxes(positions)) is False
+    assert OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions)) is False
 
 
 def test_validate_on_relations_child_z_within_tolerance_above_clearance_passes():
@@ -259,7 +253,7 @@ def test_validate_on_relations_child_z_within_tolerance_above_clearance_passes()
     box.add_relation(On(desk))
     # Child bottom 0.063 → box center z = 0.063 + 0.1 = 0.163.
     positions = {desk: (0.0, 0.0, 0.0), box: (0.0, 0.0, 0.163)}
-    assert placer._validate_on_relations(positions, _env_bboxes(positions)) is True
+    assert OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_validate_on_relations_child_z_at_or_below_parent_top_fails():
@@ -270,7 +264,7 @@ def test_validate_on_relations_child_z_at_or_below_parent_top_fails():
     box = _make_box("box", size=0.2)
     box.add_relation(On(desk))  # clearance_m=0.01; desk top 0.05
     positions = {desk: (0.0, 0.0, 0.0), box: (0.0, 0.0, 0.15)}
-    assert placer._validate_on_relations(positions, _env_bboxes(positions)) is False
+    assert OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions)) is False
 
 
 def test_on_relation_check_child_outside_xy_returns_false():
@@ -280,7 +274,7 @@ def test_on_relation_check_child_outside_xy_returns_false():
     box = _make_box("box", size=0.2)
     box.add_relation(On(desk))
     positions = {desk: (0.0, 0.0, 0.0), box: (10.0, 10.0, 0.1)}
-    assert placer._validate_on_relations(positions, _env_bboxes(positions)) is False
+    assert OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions)) is False
 
 
 def _validate_box_on_desk(edge_margin_m: float, box_x: float) -> bool:
@@ -290,7 +284,7 @@ def _validate_box_on_desk(edge_margin_m: float, box_x: float) -> bool:
     box = _make_box("box", size=0.2)
     box.add_relation(On(desk, edge_margin_m=edge_margin_m))
     positions = {desk: (0.0, 0.0, 0.0), box: (box_x, 0.0, 0.16)}
-    return placer._validate_on_relations(positions, _env_bboxes(positions))
+    return OnRelationValidator(placer.params)._validate(positions, _env_bboxes(positions))
 
 
 def test_on_relation_edge_margin_within_inset_band_passes():
@@ -319,7 +313,7 @@ def test_next_to_satisfied_passes():
     child = _make_box("child", size=0.2)
     child.add_relation(NextTo(parent, distance_m=0.05, side=Side.POSITIVE_X, cross_position_ratio=0.0))
     positions = {parent: (0.0, 0.0, 0.0), child: (0.35, 0.0, 0.0)}
-    assert placer._validate_next_to_relations(positions, _env_bboxes(positions)) is True
+    assert NextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_next_to_wrong_offset_fails():
@@ -329,7 +323,7 @@ def test_next_to_wrong_offset_fails():
     child.add_relation(NextTo(parent, distance_m=0.05, side=Side.POSITIVE_X, cross_position_ratio=0.0))
     # x=0.45 → distance off by 0.10, well past the default 0.01 tolerance.
     positions = {parent: (0.0, 0.0, 0.0), child: (0.45, 0.0, 0.0)}
-    assert placer._validate_next_to_relations(positions, _env_bboxes(positions)) is False
+    assert NextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is False
 
 
 def test_next_to_tolerance_is_per_relation():
@@ -340,7 +334,7 @@ def test_next_to_tolerance_is_per_relation():
     child = _make_box("child", size=0.2)
     child.add_relation(NextTo(parent, distance_m=0.05, side=Side.POSITIVE_X, tolerance_m=0.2))
     positions = {parent: (0.0, 0.0, 0.0), child: (0.45, 0.0, 0.0)}
-    assert placer._validate_next_to_relations(positions, _env_bboxes(positions)) is True
+    assert NextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_next_to_cross_position_not_gated_passes():
@@ -350,7 +344,7 @@ def test_next_to_cross_position_not_gated_passes():
     child = _make_box("child", size=0.2)
     child.add_relation(NextTo(parent, distance_m=0.05, side=Side.POSITIVE_X, cross_position_ratio=0.0))
     positions = {parent: (0.0, 0.0, 0.0), child: (0.35, 0.2, 0.0)}
-    assert placer._validate_next_to_relations(positions, _env_bboxes(positions)) is True
+    assert NextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 # --- NotNextTo validation (parent box XY in [-0.2, 0.2]; default keep-out margin_m = 0.1) ---
@@ -363,7 +357,7 @@ def test_not_next_to_inside_zone_fails():
     child = _make_box("child", size=0.2)
     child.add_relation(NotNextTo(parent, side=Side.POSITIVE_X))
     positions = {parent: (0.0, 0.0, 0.0), child: (0.25, 0.0, 0.0)}
-    assert placer._validate_not_next_to_relations(positions, _env_bboxes(positions)) is False
+    assert NotNextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is False
 
 
 def test_not_next_to_crossed_back_over_edge_passes():
@@ -373,7 +367,7 @@ def test_not_next_to_crossed_back_over_edge_passes():
     child.add_relation(NotNextTo(parent, side=Side.POSITIVE_X))
     # Far on the -X side: cleared the keep-out via the edge route.
     positions = {parent: (0.0, 0.0, 0.0), child: (-0.5, 0.0, 0.0)}
-    assert placer._validate_not_next_to_relations(positions, _env_bboxes(positions)) is True
+    assert NotNextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_not_next_to_slid_off_footprint_passes():
@@ -383,7 +377,7 @@ def test_not_next_to_slid_off_footprint_passes():
     child.add_relation(NotNextTo(parent, side=Side.POSITIVE_X))
     # Past the +X edge but slid well past the +Y footprint end: cleared via the cross route.
     positions = {parent: (0.0, 0.0, 0.0), child: (0.25, 0.5, 0.0)}
-    assert placer._validate_not_next_to_relations(positions, _env_bboxes(positions)) is True
+    assert NotNextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_not_next_to_tolerance_is_per_relation():
@@ -393,7 +387,7 @@ def test_not_next_to_tolerance_is_per_relation():
     child = _make_box("child", size=0.2)
     child.add_relation(NotNextTo(parent, side=Side.POSITIVE_X, tolerance_m=0.2))
     positions = {parent: (0.0, 0.0, 0.0), child: (0.25, 0.0, 0.0)}
-    assert placer._validate_not_next_to_relations(positions, _env_bboxes(positions)) is True
+    assert NotNextToValidator(placer.params)._validate(positions, _env_bboxes(positions)) is True
 
 
 def test_validate_placement_rejects_not_next_to_violation():
@@ -404,7 +398,7 @@ def test_validate_placement_rejects_not_next_to_violation():
     child.add_relation(NotNextTo(parent, side=Side.POSITIVE_X))
     # In the keep-out zone in XY, but lifted in Z so NO_OVERLAP and ON_RELATION still pass.
     positions = {parent: (0.0, 0.0, 0.0), child: (0.25, 0.0, 5.0)}
-    results = placer._validate_placement(positions, _env_bboxes(positions))
+    results = _validate_one(placer, positions, _env_bboxes(positions))
     assert results.do_all_required_validation_checks_pass() is False
     assert PlacementCheck.NOT_NEXT_TO in results.get_failed_validation_check_names
 
@@ -417,6 +411,6 @@ def test_validate_placement_rejects_next_to_violation():
     child.add_relation(NextTo(parent, distance_m=0.05, side=Side.POSITIVE_X, cross_position_ratio=0.0))
     # Offset wrong by 0.10, but lifted in Z so NO_OVERLAP and ON_RELATION still pass.
     positions = {parent: (0.0, 0.0, 0.0), child: (0.45, 0.0, 5.0)}
-    results = placer._validate_placement(positions, _env_bboxes(positions))
+    results = _validate_one(placer, positions, _env_bboxes(positions))
     assert results.do_all_required_validation_checks_pass() is False
     assert PlacementCheck.NEXT_TO in results.get_failed_validation_check_names
